gradle settings, MVP structure begining
