+++ app/build.gradle
@@ -1,31 +1,51 @@
 apply plugin: 'com.android.application'
+//apply plugin: 'realm-android'
+
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.melchuk.mygitmanager"
         minSdkVersion 23
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+
     }
+
+    compileOptions {
+        sourceCompatibility JavaVersion . VERSION_1_8
+        targetCompatibility JavaVersion . VERSION_1_8
+    }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
+
+}
+
+buildscript{
+    ext.android_support_ver  = "28.0.0"
+    ext.moxy_ver = "1.5.5"
+    ext.butterknife = "8.8.1"
+    ext.glide = "4.8.0"
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
+    implementation "com.android.support:appcompat-v7:$android_support_ver"
+    implementation "com.android.support:design:$android_support_ver"
+    implementation "com.android.support:recyclerview-v7:$android_support_ver"
+    implementation "com.android.support:cardview-v7:$android_support_ver"
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'io.reactivex.rxjava2:rxjava:2.1.0'
 }
