+++ app/src/main/java/com/melchuk/mygitmanager/data/model/Repository.java
@@ -1,0 +1,19 @@
+package com.melchuk.mygitmanager.data.model;
+
+public class Repository {
+    public final String id;
+    public final String content;
+    public final String details;
+
+    public Repository(String id, String content, String details) {
+        this.id = id;
+        this.content = content;
+        this.details = details;
+    }
+
+    @Override
+    public String toString() {
+        return content;
+    }
+
+}
