+++ app/src/main/java/com/melchuk/mygitmanager/view/repos/ReposItemFragmentRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.melchuk.mygitmanager;
+package com.melchuk.mygitmanager.view.repos;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -6,22 +6,23 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.melchuk.mygitmanager.ReposItemFragment.OnListFragmentInteractionListener;
-import com.melchuk.mygitmanager.dummy.DummyContent.DummyItem;
+import com.melchuk.mygitmanager.R;
+import com.melchuk.mygitmanager.view.repos.ReposItemFragment.OnListFragmentInteractionListener;
+import com.melchuk.mygitmanager.data.ReposDataModel.Repository;
 
 import java.util.List;
 
 /**
- * {@link RecyclerView.Adapter} that can display a {@link DummyItem} and makes a call to the
+ * {@link RecyclerView.Adapter} that can display a {@link Repository} and makes a call to the
  * specified {@link OnListFragmentInteractionListener}.
  * TODO: Replace the implementation with code for your data type.
  */
-public class MyRepositoryRecyclerViewAdapter extends RecyclerView.Adapter<MyRepositoryRecyclerViewAdapter.ViewHolder> {
+public class ReposItemFragmentRecyclerViewAdapter extends RecyclerView.Adapter<ReposItemFragmentRecyclerViewAdapter.ViewHolder> {
 
-    private final List<DummyItem> mValues;
+    private final List<Repository> mValues;
     private final OnListFragmentInteractionListener mListener;
 
-    public MyRepositoryRecyclerViewAdapter(List<DummyItem> items, OnListFragmentInteractionListener listener) {
+    public ReposItemFragmentRecyclerViewAdapter(List<Repository> items, OnListFragmentInteractionListener listener) {
         mValues = items;
         mListener = listener;
     }
@@ -60,7 +61,7 @@ public class MyRepositoryRecyclerViewAdapter extends RecyclerView.Adapter<MyRepo
         public final View mView;
         public final TextView mIdView;
         public final TextView mContentView;
-        public DummyItem mItem;
+        public Repository mItem;
 
         public ViewHolder(View view) {
             super(view);
