+++ app/src/main/java/com/melchuk/mygitmanager/view/user/UserDataFragment.java
@@ -1,4 +1,4 @@
-package com.melchuk.mygitmanager;
+package com.melchuk.mygitmanager.view.user;
 
 import android.content.Context;
 import android.net.Uri;
@@ -7,6 +7,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.melchuk.mygitmanager.R;
 
 
 /**
