+++ app/src/main/res/layout/fragment_repository_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".ReposItemFragment"
+    tools:context=".view.repos.ReposItemFragment"
     tools:listitem="@layout/fragment_repository" />
